pyproject.toml
[tool.poetry]
name = "simple-request-handler"
version = "0.1.0"
description = "Simple Python application that works as a request handler"
authors = ["Robert Hartings <robert@hartings.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.72.0"
uvicorn = "^0.17.0"
python-multipart = "^0.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
SQLAlchemy = "^1.4.29"
redis = "^4.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
